+++ CHANGELOG.md
@@ -37,6 +37,7 @@ This is a maintenance release with a few minor bugfixes.
 
 ### Added
 
+- [Gradle configuration cache](https://docs.gradle.org/current/userguide/configuration_cache.html) support
 - New [Local TS consumer setup](./samples/local-ts-consumer/README.md) sample
 
 ### Changed